Move emitter implementations to their own “emitters” folder
