+++ .gitignore
@@ -1,2 +1,4 @@
 spec/examples.txt
 *.gem
+.idea
+.DS_Store
