+++ lib/event_emitter/emitters/generic.rb
@@ -1,4 +1,4 @@
-class GenericEmitter
+class Generic
   def self.publish(*)
     fail NotImplementedError
   end
