--- lib/event_emitter/rabbitmq_emitter.rb
@@ -1,13 +1,0 @@
-require "sneakers"
-
-class RabbitMQEmitter < GenericEmitter
-
-  def initialize(options: {})
-    @publisher = Sneakers::Publisher.new(options)
-  end
-
-  def publish(message:, options:)
-    @publisher.publish(message, options)
-  end
-
-end
